+++ src/cmake/clang_format.cmake
@@ -1,11 +1,8 @@
-# Copyright 2020 Michele Adduci <adduci@tutanota.com>
-
 find_program(E57_CLANG_FORMAT_BIN NAMES clang-format)
 
 if(E57_CLANG_FORMAT_BIN)
   get_target_property(openE57_files ${PROJECT_NAME} SOURCES)
-  get_target_property(LASReader_files LASReader SOURCES)
-  get_target_property(time_conversion_files time_conversion SOURCES)
+  get_target_property(openE57las_files ${PROJECT_NAME}las SOURCES)
 
   add_custom_target(
     format-openE57
@@ -13,17 +10,12 @@ if(E57_CLANG_FORMAT_BIN)
     COMMAND_EXPAND_LISTS VERBATIM)
 
   add_custom_target(
-    format-lasreader
-    COMMAND clang-format --style=file -i ${LASReader_files}
-    COMMAND_EXPAND_LISTS VERBATIM)
-
-  add_custom_target(
-    format-timeconv
-    COMMAND clang-format --style=file -i ${time_conversion_files}
+    format-openE57las
+    COMMAND clang-format --style=file -i ${openE57las_files}
     COMMAND_EXPAND_LISTS VERBATIM)
 
   add_custom_target(
     format
     COMMENT "Running clang-format..."
-    DEPENDS format-openE57 format-lasreader format-timeconv)
+    DEPENDS format-openE57 format-openE57las)
 endif()