+++ CHANGELOG.md
@@ -5,12 +5,13 @@
 ## Changed
 - Added conan packaging support
 - Refactored project structure
-- Renamed LASReader to openE57las 
+- Renamed LASReader to openE57las
 - Moved time_conversion to tools folder
 
 ## Removed
 - config.h template (definition passed through the compiler)
 - Removed the dependency of time_conversion from openE57
+- Removed the generation of time_conversion as artifact of openE57 (used only by tools)
 
 ## [1.5.1] - 2021-12-09