+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Refactored project structure
 - Renamed LASReader to openE57las
 - Moved time_conversion to tools folder
+- Updated Boost to 1.78.0 (only used if building tools)
 
 ## Removed
 - config.h template (definition passed through the compiler)